% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessTrace.R
\name{createTimedAutomaton}
\alias{createTimedAutomaton}
\title{Create Automaton from timed-regex}
\usage{
createTimedAutomaton(timedRegEx)
}
\arguments{
\item{create}{Timed-automaton from timed-regular expressions}
}
\value{
Nothing
}
\description{
Create Automaton from timed-regex
}

