% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessTrace.R
\name{getMinedSpecifications}
\alias{getMinedSpecifications}
\title{Find a set of sequences of traceEvents which match the given timed regular expression
and the required probability threshold}
\usage{
getMinedSpecifications(resultList, threshold = 0.9)
}
\arguments{
\item{threshold}{probability to select mined timed-regular expressions}

\item{a}{List returned by processTrace}
}
\value{
List whose elements represent mined events
}
\description{
Find a set of sequences of traceEvents which match the given timed regular expression
and the required probability threshold
}

