% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessTrace.R
\name{processTrace}
\alias{processTrace}
\title{Process timed regex}
\usage{
processTrace(traceTimes, traceEvents, alphabetLength, timedRegEx)
}
\arguments{
\item{traceTimes}{Numeric vector of timestamps.}

\item{traceEvents}{Integer vector of corresponding events.}

\item{alphabetLength}{Integer greater than zero indicating number of unique events.}

\item{timedRegEx}{the timed-regular expression template}
}
\value{
List of two nd-arrays.
Success and Reset are counters values for each alphabet configuration.
}
\description{
Processes the trace on a given timed regex, returning number of successes and resets
}

